Prefer balance in payment DB when balance from wallet is 0
This prevents an issue at startup where electrum returns a zero balance
because it is too slow to fetch the relevant data from the servers, even
if eclair wallet knows from its DB that the balance may not be 0.

In last resort we rely on the onchain txs recorded in the Eclair Wallet
payment DB.

If the balance from the wallet is truly 0 and the balance from the payment DB
returns a wrong non zero balance, it can only be because Eclair Wallet was
not notified with the latest txs, and should be notified soon, in which
case the balance from the payment DB will be corrected.