Onchain balance is centralized in atomic reference in App
`WalletBalanceUpdateEvent` should not be used to store the current onchain
balance because that's a hack (see onCreate in HomeActivity) and the stored
value is redundant with the balance already stored in `App`'s
`onChainBalance` atomic reference. This atomic reference is now the unique
source of truth.

`WalletBalanceUpdateEvent` is now simply used as a ping
event to refresh the gui.