Using max funding limits from eclair-core
The minimal funding amount is 1 mBTC which is an arbitrary and conservative
amount. We can not use the `MIN_FUNDING_SATOSHIS` from `Channels` in
eclair core because it does not include the fees and as such is useless to
the user. A funding with such a low amount would automatically fail
because the fees would make the funding amount less than the dust limit.

We now display the funding limits in the user's prefered unit

Also removed hardcoded maximum limit from the ui.