Saving last known exchange rates in preferences
When the wallet starts, if the exchange rate is not persisted and the
price api is slow to answer, amounts can not be converted to fiat. We now
save the last known rates so that amounts can always be converted to fiat.

The drawback is that the last known price can be very out of date and give
the user a wrong impression about his fiat holdings.