Display a floating message when electrum is desynchronized
By using the latest block's timestamp in the ElectrumReady message, the
wallet can autonomously estimate the freshness of the electrum server
it's connected with and warn the user. If the user is lagging behind
the actual network, its peers may unilaterally close the channels with him
if he attempts payments. The user should be warned about this.