Using CoinUtils from eclair-core for money conversions
We should use the utility from eclair-core to convert bitcoin amount to
strings, and vice-versa. Static methods for fiat and invoice reading are
moved to WalletUtils, which is more logical anyway.

The new CoinUtils is now able to display Satoshi amounts with decimals.
The millisatoshi part when showing amounts in satoshi is not rounded anymore.

Added a user preference for the number of decimals that should be
displayed. Lightning sometimes handles very small amounts which make for
poor readibility. The user should be able to filter some decimals if he
wants to.