Payment result handling is now detached from payment execution
Core now dispatches a payment result object when a payment fails or
succeeds. This event is listened to in eclair event service.

In the android app context, a payment must first be in the INIT status,
then PENDING once a ChannelSignatureSent event with a corresponding htlc is
emitted, then PAID or FAILED according to the result.