Added 'empty wallet' feature
This feature helps to clean the wallet of any bitcoin funds by building
a transaction with amount and network fees such that all the available
funds are spent and that no dust remains in the wallet.

It's the only case where fees are included in the displayed amount (this
is implicit).