Seed is now encrypted with a 6 digits passwords (#2)
After creating a new wallet or importing an existing one, the seed is
encrypted with java-aes-crypto and written to disk.

The seed must be decrypted first before the application can be started. Once
decrypted the password is kept in memory in case the user has activated
the check pin before payment.

If an unencrypted seed is present in the app data directory, it is
encrypted and written to disk before the application can be used.

Added an EncryptedSeed object that handle serialization to byte array
and deserialization from byte array to EncryptedSeed, and handles said
encryption.