Control of active channels is done when actually sending payments
Checking the state of the channels too early in the payment process may
trigger false positives, especially when opening an invoice from an
intent filter. The node may have active channels but they may be in
a transient state because the app just started, and the error would
show. This scenario is especially true on faster phones.