Added support for more fiat currencies
Also reworked the way the fiat exchange rate is retrieved. A static
object containing the exchange rates is still necessary (due to the fact
that SharedPreferences - which is already storing the exchange rates -
requires access to the application context, which is not static).

Switched to the blockchain.info rate api which offers many currencies
conversion within a single call.

fixes #71