Added extra routing info when generating payment request
Channels to the wallet are private and extra routing information must be
added to any payment request generated by the wallet. Otherwise, the
payer will not be able to find a route to the wallet and payments will
fail.

Extra routes are local channels in normal state. If no channels are
available, payments will not suceed.