Local channels are now saved in application database
Added a `LocalChannel` object that represents the channel in the
database. Some of its properties are transient and are not saved in DB.
This object replaces the `ChannelDetails` object in
`EclairEventService`.

A channel can be active (currently used by the node, event if its state
is closing or closed) or inactive (channel was closed and forgotten by
the node).

Saving inactive channels in DB will enable several features such as
linking a payment to a channel or tracking usage).

Inactive channels can be displayed in the local channels list.

The local channel item look has been updated, and a progress bar
showing the channel balance over the channel capacity has been
added.