Added a Backup mechanism for active lightning channels
Channels can now be saved in google drive. An event `ChannelPersisted`
is now dispatched by the node when the channels database is updated.
On event reception, a backup is sent to google drive (with a limit of
one per sec). If the user does not have google drive or does not grant
access to the application, the feature is not available and the app
works as before.

If the user loses access to his phone, he can then restore all the
active channels from this node, provided the feature was activated and
the application was given access when the backup was to be made, and
the user chose to restore the backup when he starts the application
with the same seed.

A backup is attached to a seed through the hash of the wallet publick key.
Several backup can coexist on the user's drive. Backup files are stored
in the Drive application folder (invisible to the user, which prevents
any hazardous editing).