Added specific class for enabling channels backup on first start
This streamlines the first start experience. Channels backup settings is
now managed from a specific page with the same look and feel as other
settings pages.

The restore channels feature now request synchronization from drive to
make sure that the app has a up-to-date view of the Drive application
folder. Note that this fixes a bug where, after uninstalling app, Drive
removes the local content from the app folder, and once the user
reinstalls the app, no channel backups can be found because no
synchronization with the server is made.