Channels backup is now handled by the WorkManager API
The WorkManager framework simplifies the backup scheduling process. The
backup is handled in an background task detached from the application,
and is guaranteed to be run by the system.

Removed Backup Event, Backup scheduler, Backup intent service which are
replaced by the WorkManager api. Also removed the termination service
that would trigger a backup when the application closes. This is not
needed anymore.