Added a disclaimer when restoring a backup created by another device
Added the id of the device in the backup metadata. With this id, the app
can tell whether the existing backup in Drive comes from the same device,
or not.

We use the ANDROID_ID as identifier instead of the instance id because
we need an id attached to the device. Otherwise the user would get false
negatives when reinstalling the app from the same device.

Thanks to this id, the app can display a disclaimer warning the user
intends not to use the same channels from two different applications
because his channels would eventually be unilaterally closed, which is
costly.