Routing Sync Updates (#117)
Use ACINQ node to sync our routing table.

Local features are set to data-loss-protect and nothing else.
When we start, we automatically open a connection to the ACINQ node and a start
syncing our routing table. A connection is also scheduled every minute - and when the network connectivity is reestablished.


Also:

* Network connectivity is checked at startup

Node bootstrap requires network access to complete, which block
application startup is the device has no connectivity to the internet.

An error is now displayed ASAP, instead of letting the user wait for
60s and then fail. The next step is to decouple the startup from the
bootstrap and letting the app safely starts in offline mode.

* Removed obsolete DB compatibility checks.

* Throttled UI refresh events to improve perfs

UI Events dispatched from eclair/electrum watcher services must be
throttled to limit the rate of UI redraws, especially the payments and
the channels list which can grow quite large.