Reworked wallet creation activity
Words are displayed in a more logical fashion (columns instead
of rows).

In the verification process, each one of the 3 words has its own
input, instead on one input which is splitted around a ` ` space
character, which was error prone and induced false negatives.

Also added more informative feedback messages.