Use ViewPager for seed creation proces and added passphrase
User can also add a passphrase to a seed when creating a wallet.
Passphrase is optional, and must be confirmed when it is not blank.

Seed verification layout is improved by using separate fields for each
words that the user must verify. The layout is animated when the
verification fails.