Wallet creation/import overhaul (#121)
When creating a seed, the recovery words are displayed in a more
logical fashion (columns instead of rows). In the verification process, 
each one of the 3 words has its own input, instead on one input 
splitted around a ` ` space character, which was error prone and 
induced false negatives. The layout is animated when the verification 
fails.

Design wise, we use ViewPager for seed creation and seed import,
which improve UX by adding swipe animations, and back button 
support.

User can now add a passphrase to a seed when creating or restoring
a seed. This passphrase is optional, and must be confirmed when it is 
not blank.