Fixed scrolling in payment window and improved custom fee display
Fixed an issue where displaying the keyboard would conflict with the
scroll view in the send payment modal window.

The fee button changes its look when the user enters a custom value.

Added a bottom border to the fee input to make it clear that this field
can be edited.

fixes #113