Use all available funds when opening a channel
Added a checkbox in the open channel window to use all the available
on-chain funds when opening a channel. This amount cannot exceed
the maximum channel capacity (~167 mBTC to date).

Note that in some cases the calculated amount will not be accurate and
some dust can be leftover. This will be fixed in a later release of
eclair-core.

fixes #124