Added a reconnect feature for OFFLINE nodes
Some nodes have a dynamic IP which means that when the IP changes the
node's channels may go offline without a way for the app to reconnect
to this node.

This commit adds a button in offline channels page to reconnect to a
node. The user can either scan a node uri QR code (same as in open
channel) or enter an IP manually (or copy/paste). The app then sends a
Peer.Connect request to the new uri. Note that the user can only connect
to the same node (same node id).

fixes #125