Reject expired lightning payments
The wallet now enforce the expiry with a 1 hour default value if the
payment request does not define one. This is compliant with the specs:

https://github.com/lightningnetwork/lightning-rfc/blob/25a9debcf3c9895b71c1d1648bb29d9ff07d9cde/11-payment-encoding.md#tagged-fields

Also added a progress bar to the loading message displayed when the
invoice is being read.

Removed the readonly logic for the payment amount input because the user
should always be able to override the payment amount, especially with
lightning.

fixes #128