Pay button is disabled for LN payment if channels are offline
If all the channels are offline when the payment is initiated, the
payment should not be cancelled with an error message. Instead, the pay
button is disabled which has the same effect but is a better UX. A
polling logic is added to check every second if the state of the
channels has changed, in which case the button is enabled back.

This is especially useful for payments started from a cold start wallet,
where channels would not have time to establish connection and the
payment would be cancelled.

fixes #131