Added a background task synchronizing the routing table
A worker synchronizing the routing table (that is the network sqlite DB)
is scheduled when the device starts and when the app starts.

It uses a lighter setup and uses a random seed since the actual seed is
encrypted and unavailable from a background service.

This feature should improve the chance of successful payments since the
routing table should never be stale ; this will be especially effective
for users that do not start the application often.