Cleaning zero-confs txs on start
Some txs never get into the mempool and linger in the payment history.
Cleaning these txs from the payment database when starting the app
ensures that the wallet starts with a clean slate.

Regular successful zero confs txs will be will be sent again by
ElectrumWallet when starting the app.