Redesign of payment item
Removed payment status when unnecessary, that is most of the times: when
payment is successful or when it's an on-chain tx.

Removed bottom border and increased white spaces.

Payment date is now relative. The payment type has also been updated
for consistency.

The layout is now a constraint layout, which is more efficient and
improve performances.