Added a warning message if the electrum check has not run for a while
Some devices have an aggressive way of optimizing applications which
prevents any background work to be run. We verify electrum check
timestamps to know if the background electrum check has run recently.
If not, a warning is displayed.

Warning contains a link to an external website dealing about this issue
and containing some useful instructions for userss.