Added timestamp for electrum check worker attempts
This fixes an issue where the background warning message would display
when not needed. This new timestamp fixes that, with one trade-off: the
message will not be displayed if the user regularly starts the app, even
though it should. This however should not be a problem since, in this
scenario, the app runs regularly enough so that a background electrum
check is in fact not necessary.

This commit also upgrade the WorkManager library to 1.0.0.