Migrate from BinaryData to ByteVector (#165)
Upgraded to the latest eclair-core version which removes the `BinaryData` object (see
pull requests ACINQ/eclair#896 and ACINQ/bitcoin-lib#31). We should now use byte 
arrays (and affiliated objects) instead.

Some android tests have been fixed. An issue remains with the ElectrumClient test, 
where it does not seem possible to easily instantiate SSL parameters from Java.

The seed is now correctly read. It was saved as an hex string and not as a binary file.