Wallet readiness when sending payment now checks block height
Wallet readiness check now accounts for wallet's current block height,
and for electrum connection state.

If the wallet blockheight is unreasonably low, it is safe to assume that
the wallet is not properly synced and payment should not be sent. This
prevents an issue where a user could send a LN payment with a low
expiry which, while valid, could force peers to close the relaying channel.

On-chain payments will now also check wallet readiness.