Removed breaking version case in startup and updated page layout
No version of the app should ever break compatibility so the code has
been removed. Note that this code was only relevant for version 14 which
indeed breaks compatibility with v13, but these versions were only
released on testnet and are very old.

Also, app should use constraint layouts as much as possible.