Add support for local backup (#188)
App now requires access to local storage (write). This access is
mandatory to start the app, to make sure that there always is a local
backup.

Channel restore page has been reworked to handle both local and gdrive
backup. Retrieved files are compared based on the maximum commitment
index count, the backup with highest being automatically used as the
preferred backup. The page's flow has also been improved

If there is no access to the local storage, when starting the app, a
blocking window is displayed to ask for local storage access. Access to
gdrive is only requested if this is the user has never started the app
yet ('started' means here that the node has been started).

Channels backup settings page has also been revamped.

Note that we do not use the `ChannelPersisted` event anymore, instead we
use the proper `BackupEvent` sent by the node once the backup file has
been generated and is safe to use.