User can now enter a fiat amount when sending payment
A button has been added in the payment modal window to switch between a
fiat input and a bitcoin input.

WIP: must handle the case where conversion rates are old, which will
disable this feature to prevent a case where the user would spend money
based on obsolete data.

Note that the app will always send the bitcoin amount displayed on
the modal window.