Migrate google drive api to the REST V3 api
The Google Drive Android API that was used to upload/retrieve channels
backup has been deprecated and scheduled for termination at the end of
Dec, 2019. We switch to the REST V3 API which provides almost the same
feature set, albeit without offline syncing. The lack of offline syncing
is not an issue since we already save the channels backup on-device.

see: https://developers.google.com/drive/android/deprecation

* Hidden AppData folder

Since Google mentions that the hidden AppData folder is going to be
deprecated at some point, and since the user has already access to the
actual backup file, there is no reason to hide the backups and the
Google Drive backup files are now stored in a regular Drive folder that
the user can freely access.

Note that the app still checks for old backup that would be in this
hidden AppData folder. Subsequent backups will be uploaded to the public
folder.

* Backup compression

The backup files are now compressed before being encrypted to save
some bandwidth.

Fixes #208