Added support for biometric wallet unlock
User can now unlock the wallet by using biometric authentication
(fingerprint, iris, face...) if the device allows it. Limited to android
8+ for now, need testing for 7+.

We use the Android Keystore to encrypt the app PIN, and then decrypt the
PIN when user is biometrically authenticated. User can always fallback
to PIN if needed.

Adds Kotlin support to the project.