Use legacy mode for external storage access
Android 10 (api 29) deprecates the external storage methods, and instead
recommend using the getExternalDir methods to store app-specific files.
However for backups this is not good enought since these files are
removed when the application is uninstalled.

The legacy option lets the app use the previous API while a better
solution is found. Note that with Android 11 (R) which is not yet
released, this flag will not work.

A possible solution could be to use the Shared storage with the
MediaStore.Files API, however this would requires additionals
permissions: MANAGE_EXTERNAL_STORAGE +
ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION