Enable basic mpp
This commits enables the basic_mpp feature bit.

The UI controls checking if an outgoing payment can be sent now use the
aggregated balance of all channels, instead of the max balance of all
channels. The inbound liquidity warning still use the max inbound
balance and not the aggregate.

This does not mean that the app is able to send/receive the full amount
in one payment. It also depends on the other nodes in then network and
how well they support mpp.