Use MediaStore API to store local backup on Android 10+
Du to changes in Android 10+ regarding shared storage files, we must use
the Media Store API to save/read the channels backup files in a shared
folder (which are not deleted when app is uninstalled).

We also use the Storage Access Framework to allow the user to manually
select a backup using the device's file browser. This is necessary in
case the backup is added manually, in which case the file would not be
visible by the app when using the Media Store API (the file would not be
created by the app).

As such the storage write permission is not required anymore for Android 10+.